Add new fiducial_positions, transformed using AG measurements.
We assume that the fiducial fiber and guide camera positions as
measured by the XY stage at ASRD are good relative to each other, but
not good relative to the PFI frame.

Now that we have on-sky AG measurements, transform all positions to
link sky to PFI mm.

Specifically, as listed in INSTRM-1706:

 dX = -1.748mm
 dY = -0.963mm
 dT = +0.684171deg
 dS = +1.11458e-04  (i.e., x 1.000111458)
(xy scanner -> sky images)